Correção do PUT e melhorias de orientacao a objeto
Alterações nos pacotes, inclusão de construtores, Inclusão de validação de campos obrigatórios, ajuste de PUT para impedir alteração de saldo ao editar consumidor

+++ src/main/java/br/com/alelo/consumer/consumerpat/respository/ConsumerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import br.com.alelo.consumer.consumerpat.model.Consumer;
 
-public interface ConsumerRepository extends JpaRepository<Consumer, Integer> {
+public interface ConsumerRepository extends JpaRepository<Consumer, Long> {
 
 //    @Query(nativeQuery = true, value = "select * from Consumer")
 //    List<Consumer> getAllConsumersList();
